# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Hush API
  version: 1.0.0
  description: >
    Contract for Hush Apple clients, Agent backend, and Unified Inbox.
    Quest steps are canonical in the fixed content library; the backend primarily returns quest_id.
servers:
  - url: https://api.example.com
    description: Demo
  - url: http://localhost:3000
    description: Local
...
paths:
...
  /v1/inbox/events:batch:
    post:
      tags: [Inbox]
      operationId: ingestInboxEvents
      security:
        - ConnectorBearer: []
        - DemoTokenAuth: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/ClientVersion"
        - $ref: "#/components/parameters/ContractVersion"
        - $ref: "#/components/parameters/DemoToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./schemas/inbox-event-batch.schema.json"
      responses:
        "202":
          description: Events accepted after deduplication
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
            X-Contract-Version:
              $ref: "#/components/headers/ContractVersion"
            X-Hush-Data-Origin:
              $ref: "#/components/headers/DataOrigin"
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [accepted, duplicates, item_ids]
                properties:
                  accepted: { type: integer, minimum: 0 }
                  duplicates: { type: integer, minimum: 0 }
                  item_ids:
                    type: array
                    items: { type: string }
        "400":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "503":
          $ref: "#/components/responses/Error"
  /v1/inbox/items:
    get:
      tags: [Inbox]
      operationId: listInboxItems
      security:
        - AppBearer: []
        - DemoTokenAuth: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/ClientVersion"
        - $ref: "#/components/parameters/ContractVersion"
        - $ref: "#/components/parameters/DemoToken"
        - $ref: "#/components/parameters/AppSession"
        - name: cursor
          in: query
          required: false
          schema: { type: string }
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        "200":
          description: Unified Inbox page
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
            X-Contract-Version:
              $ref: "#/components/headers/ContractVersion"
            X-Hush-Data-Origin:
              $ref: "#/components/headers/DataOrigin"
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items, next_cursor]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "./schemas/inbox-item.schema.json"
                  next_cursor:
                    type: [string, "null"]
        "400":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /v1/inbox/items/{itemId}:
    get:
      tags: [Inbox]
      operationId: getInboxItem
      security:
        - AppBearer: []
        - DemoTokenAuth: []
      parameters:
        - name: itemId
          in: path
          required: true
          schema: { type: string }
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/ClientVersion"
        - $ref: "#/components/parameters/ContractVersion"
        - $ref: "#/components/parameters/DemoToken"
        - $ref: "#/components/parameters/AppSession"
      responses:
        "200":
          description: Unified Inbox item
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
            X-Contract-Version:
              $ref: "#/components/headers/ContractVersion"
            X-Hush-Data-Origin:
              $ref: "#/components/headers/DataOrigin"
          content:
            application/json:
              schema:
                $ref: "./schemas/inbox-item.schema.json"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /v1/inbox/items/{itemId}/summary:
    post:
      tags: [Inbox]
      operationId: summarizeInboxItem
      security:
        - AppBearer: []
        - DemoTokenAuth: []
      parameters:
        - name: itemId
          in: path
          required: true
          schema: { type: string }
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/ClientVersion"
        - $ref: "#/components/parameters/ContractVersion"
        - $ref: "#/components/parameters/DemoToken"
        - $ref: "#/components/parameters/AppSession"
      responses:
        "200":
          description: Item with refreshed AI summary
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
            X-Contract-Version:
              $ref: "#/components/headers/ContractVersion"
            X-Hush-Data-Origin:
              $ref: "#/components/headers/DataOrigin"
          content:
            application/json:
              schema:
                $ref: "./schemas/inbox-item.schema.json"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "503":
          $ref: "#/components/responses/Error"
  /v1/inbox/items/{itemId}/draft:
    post:
      tags: [Inbox]
      operationId: createInboxDraft
      security:
        - AppBearer: []
        - DemoTokenAuth: []
      parameters:
        - name: itemId
          in: path
          required: true
          schema: { type: string }
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/ClientVersion"
        - $ref: "#/components/parameters/ContractVersion"
        - $ref: "#/components/parameters/DemoToken"
        - $ref: "#/components/parameters/AppSession"
      responses:
        "200":
          description: Existing or generated reply draft
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
            X-Contract-Version:
              $ref: "#/components/headers/ContractVersion"
            X-Hush-Data-Origin:
              $ref: "#/components/headers/DataOrigin"
          content:
            application/json:
              schema:
                $ref: "./schemas/inbox-draft.schema.json"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "503":
          $ref: "#/components/responses/Error"
  /v1/inbox/drafts/{draftId}:
    get:
      tags: [Inbox]
      operationId: getInboxDraft
      security:
        - AppBearer: []
        - DemoTokenAuth: []
      parameters:
        - name: draftId
          in: path
          required: true
          schema: { type: string }
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/ClientVersion"
        - $ref: "#/components/parameters/ContractVersion"
        - $ref: "#/components/parameters/DemoToken"
        - $ref: "#/components/parameters/AppSession"
      responses:
        "200":
          description: Reply draft
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
            X-Contract-Version:
              $ref: "#/components/headers/ContractVersion"
            X-Hush-Data-Origin:
              $ref: "#/components/headers/DataOrigin"
          content:
            application/json:
              schema:
                $ref: "./schemas/inbox-draft.schema.json"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
    patch:
      tags: [Inbox]
      operationId: updateInboxDraft
      security:
        - AppBearer: []
        - DemoTokenAuth: []
      parameters:
        - name: draftId
          in: path
          required: true
          schema: { type: string }
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/ClientVersion"
        - $ref: "#/components/parameters/ContractVersion"
        - $ref: "#/components/parameters/DemoToken"
        - $ref: "#/components/parameters/AppSession"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./schemas/inbox-draft.schema.json#/$defs/InboxDraftPatch"
      responses:
        "200":
          description: User-edited reply draft
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
            X-Contract-Version:
              $ref: "#/components/headers/ContractVersion"
            X-Hush-Data-Origin:
              $ref: "#/components/headers/DataOrigin"
          content:
            application/json:
              schema:
                $ref: "./schemas/inbox-draft.schema.json"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /v1/inbox/drafts/{draftId}/confirmation:
    post:
      tags: [Inbox]
      operationId: confirmInboxDraft
      security:
        - AppBearer: []
        - DemoTokenAuth: []
      parameters:
        - name: draftId
          in: path
          required: true
          schema: { type: string }
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/ClientVersion"
        - $ref: "#/components/parameters/ContractVersion"
        - $ref: "#/components/parameters/DemoToken"
        - $ref: "#/components/parameters/AppSession"
      responses:
        "200":
          description: Short-lived one-time user confirmation
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
            X-Contract-Version:
              $ref: "#/components/headers/ContractVersion"
            X-Hush-Data-Origin:
              $ref: "#/components/headers/DataOrigin"
          content:
            application/json:
              schema:
                $ref: "./schemas/inbox-send.schema.json#/$defs/InboxConfirmation"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
...
  /v1/inbox/sync-status:
    get:
      tags: [Inbox]
      operationId: getInboxSyncStatus
      security:
        - AppBearer: []
        - DemoTokenAuth: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/ClientVersion"
        - $ref: "#/components/parameters/ContractVersion"
        - $ref: "#/components/parameters/DemoToken"
        - $ref: "#/components/parameters/AppSession"
      responses:
        "200":
          description: Per-account provider coverage and checkpoint state
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
            X-Contract-Version:
              $ref: "#/components/headers/ContractVersion"
            X-Hush-Data-Origin:
              $ref: "#/components/headers/DataOrigin"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "./schemas/inbox-sync-status.schema.json"
        "409":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    AppBearer:
      type: http
      scheme: bearer
      description: Hush App session token. Cannot ingest connector events.
    ConnectorBearer:
      type: http
      scheme: bearer
      description: Connector Host token. Cannot query items or send drafts.
    DemoTokenAuth:
      type: apiKey
      in: header
      name: X-Hush-Demo-Token
      description: Available only when HUSH_DEMO_MODE=true.
  parameters:
    RequestId:
      name: X-Request-ID
      in: header
      required: true
      description: Must equal request_id in JSON request bodies.
      schema:
        type: string
        minLength: 1
    ClientVersion:
      name: X-Client-Version
      in: header
      required: true
      schema:
        type: string
        minLength: 1
    ContractVersion:
      name: X-Contract-Version
      in: header
      required: true
      schema:
        type: string
        const: "1.0"
    DemoToken:
      name: X-Hush-Demo-Token
      in: header
      required: false
      description: Enables Sample Mode only when the server also has HUSH_DEMO_MODE=true.
      schema:
        type: string
        minLength: 8
    AppSession:
      name: X-Hush-App-Session
      in: header
      required: true
      description: >
        Per-launch high-entropy App session secret. Confirmation tokens are
        bound to its authenticated digest.
      schema:
        type: string
        minLength: 32
        maxLength: 128
...
  headers:
    RequestId:
      schema:
        type: string
    ContractVersion:
      schema:
        type: string
        const: "1.0"
    DataOrigin:
      schema:
        type: string
        enum: [real, mock, cached]
  responses:
    Error:
      description: Error
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
        X-Contract-Version:
          $ref: "#/components/headers/ContractVersion"
        X-Hush-Data-Origin:
          $ref: "#/components/headers/DataOrigin"
      content:
        application/json:
          schema:
            $ref: "./schemas/error-response.schema.json"